models.py was splitted into smaller files
