+++ models/__init__.py
@@ -1,0 +1,4 @@
+from .Kmean import KMeans
+from .KNearestNeighbors import KNearestNeighbors
+from .LogisticRegression import LogisticRegression
+from .NeuralNet import NeuralNet
